requirements.txt
openai==0.28.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3 
flask
flask-cors